refactor(pages): rename folder from tmp to home
